+++ src/app/live/live.component.ts
@@ -9,12 +9,12 @@ import { PonyComponent } from '../pony/pony.component';
 import { EMPTY, Subject, bufferToggle, catchError, filter, groupBy, interval, map, mergeMap, switchMap, throttleTime } from 'rxjs';
 import { FINISHED_RACE_STATUS, RUNNING_RACE_STATUS } from 'src/constants/status';
 import { FromNowPipe } from '../from-now.pipe';
-import { AlertComponent } from '../alert/alert.component';
+import { NgbAlert } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'pr-live',
   standalone: true,
-  imports: [CommonModule, PonyComponent, FromNowPipe, AlertComponent],
+  imports: [CommonModule, PonyComponent, FromNowPipe, NgbAlert],
   templateUrl: './live.component.html',
   styleUrls: ['./live.component.css']
 })