+++ src/app/login/login.component.spec.ts
@@ -1,13 +1,12 @@
 import { fakeAsync, tick, TestBed } from '@angular/core/testing';
-import { provideRouter } from '@angular/router';
-import { Router } from '@angular/router';
+import { provideRouter, Router } from '@angular/router';
 import { By } from '@angular/platform-browser';
+import { NgbAlert, NgbAlertConfig } from '@ng-bootstrap/ng-bootstrap';
 import { Subject } from 'rxjs';
 
 import { LoginComponent } from './login.component';
 import { UserService } from '../user.service';
 import { UserModel } from '../models/user.model';
-import { AlertComponent } from '../alert/alert.component';
 
 describe('LoginComponent', () => {
   let userService: jasmine.SpyObj<UserService>;
@@ -17,6 +16,9 @@ describe('LoginComponent', () => {
     TestBed.configureTestingModule({
       providers: [provideRouter([]), { provide: UserService, useValue: userService }]
     });
+    // turn off the animation for the alert
+    const alertConfig = TestBed.inject(NgbAlertConfig);
+    alertConfig.animation = false;
   });
 
   it('should have a credentials field', () => {
@@ -196,15 +198,16 @@ describe('LoginComponent', () => {
     fixture.detectChanges();
 
     const element = fixture.debugElement;
-    const alert = element.query(By.directive(AlertComponent));
-    expect(alert).withContext('You should have an AlertComponent to display an error message').not.toBeNull();
+    const alert = element.query(By.directive(NgbAlert));
+    expect(alert).withContext('You should have an NgbAlert to display an error message').not.toBeNull();
     expect(alert.nativeElement.textContent).toContain('Nope, try again');
-    expect(alert.componentInstance.type).withContext('The alert should be a danger one').toBe('danger');
+    const alertComponent = alert.componentInstance as NgbAlert;
+    expect(alertComponent.type).withContext('The alert should be a danger one').toBe('danger');
 
     // close the alert
-    alert.componentInstance.closeHandler();
+    alertComponent.close().subscribe();
     fixture.detectChanges();
-    expect(element.query(By.directive(AlertComponent)))
+    expect(element.query(By.directive(NgbAlert)))
       .withContext('The alert should disappear when closed')
       .toBeNull();
   });