+++ src/app/login/login.component.ts
@@ -3,15 +3,15 @@ import { NgIf } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { UserService } from '../user.service';
 import { Router } from '@angular/router';
-import { AlertComponent } from '../alert/alert.component';
 import { FormControlValidationDirective } from '../form-control-validation.directive';
 import { FormLabelDirective } from '../form-label.directive';
 import { FormLabelValidationDirective } from '../form-label-validation.directive';
+import { NgbAlert } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'pr-login',
   standalone: true,
-  imports: [NgIf, FormsModule, AlertComponent, FormControlValidationDirective, FormLabelDirective, FormLabelValidationDirective],
+  imports: [NgIf, FormsModule, NgbAlert, FormControlValidationDirective, FormLabelDirective, FormLabelValidationDirective],
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.css']
 })